refactor(server): centralize avatar url handling
